+++ src/Telegram.Bot/Requests/Sending Messages/SendMediaGroupRequest.cs
@@ -24,7 +24,7 @@ public partial class SendMediaGroupRequest() : FileRequestBase<Message[]>("sendM
 
     /// <summary>Identifier of the direct messages topic to which the messages will be sent; required if the messages are sent to a direct messages chat</summary>
     [JsonPropertyName("direct_messages_topic_id")]
-    public int? DirectMessagesTopicId { get; set; }
+    public long? DirectMessagesTopicId { get; set; }
 
     /// <summary>Sends messages <a href="https://telegram.org/blog/channels-2-0#silent-messages">silently</a>. Users will receive a notification with no sound.</summary>
     [JsonPropertyName("disable_notification")]
