+++ src/Telegram.Bot/Types/Gift.cs
@@ -15,11 +15,11 @@ public partial class Gift
     /// <summary>The number of Telegram Stars that must be paid to send the sticker</summary>
     [JsonPropertyName("star_count")]
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public int StarCount { get; set; }
+    public long StarCount { get; set; }
 
     /// <summary><em>Optional</em>. The number of Telegram Stars that must be paid to upgrade the gift to a unique one</summary>
     [JsonPropertyName("upgrade_star_count")]
-    public int? UpgradeStarCount { get; set; }
+    public long? UpgradeStarCount { get; set; }
 
     /// <summary><em>Optional</em>. The total number of the gifts of this type that can be sent; for limited gifts only</summary>
     [JsonPropertyName("total_count")]
