+++ src/Telegram.Bot/Types/GiveawayCreated.cs
@@ -6,10 +6,10 @@ public partial class GiveawayCreated
 {
     /// <summary><em>Optional</em>. The number of Telegram Stars to be split between giveaway winners; for Telegram Star giveaways only</summary>
     [JsonPropertyName("prize_star_count")]
-    public int? PrizeStarCount { get; set; }
+    public long? PrizeStarCount { get; set; }
 
-    /// <summary>Implicit conversion to int (PrizeStarCount)</summary>
-    public static implicit operator int?(GiveawayCreated self) => self.PrizeStarCount;
-    /// <summary>Implicit conversion from int (PrizeStarCount)</summary>
-    public static implicit operator GiveawayCreated(int? prizeStarCount) => new() { PrizeStarCount = prizeStarCount };
+    /// <summary>Implicit conversion to long (PrizeStarCount)</summary>
+    public static implicit operator long?(GiveawayCreated self) => self.PrizeStarCount;
+    /// <summary>Implicit conversion from long (PrizeStarCount)</summary>
+    public static implicit operator GiveawayCreated(long? prizeStarCount) => new() { PrizeStarCount = prizeStarCount };
 }
